Edit Roles Crud

+++ app/Http/Controllers/Admin/RoleController.php
@@ -7,6 +7,7 @@ use App\Http\Controllers\Controller;
 use App\Role;
 use Illuminate\Support\Facades\Validator;
 use App\Http\Requests\CreateRoleRequest;
+use App\Http\Requests\EditRoleRequest;
 
 class RoleController extends Controller
 {
@@ -66,9 +67,8 @@ class RoleController extends Controller
      * @param  int  $id
      * @return \Illuminate\Http\Response
      */
-    public function edit($id)
+    public function edit(Role $role)
     {
-        $role = Role::find($id);
         return view('administrations.roles.edit')->with('role', $role);
     }
 
@@ -77,34 +77,13 @@ class RoleController extends Controller
      *
      * @param  \Illuminate\Http\Request  $request
      * @param  int  $id
-     * @return \Illuminate\Http\Response
      */
-    public function update(CreateRoleRequest $request, Role $role)
+    public function update(EditRoleRequest $request, Role $role)
     {
-        /* CUSTOM SPECIFIC VALIDATION */
-        $customError = null;
-        // Check if the name already exists AND is not the same between the form POST and the DB
-        // This way, we can edit just the description and save the same name, but we cannot save the same name as an other sport on DB
-        if($role->slug != $request->input('slug') && Role::where('slug', '=', $request->input('slug'))->exists()){
-            $customError = 'le role "'.$request->input('slug').'"'.' existe déjà.';
-        }
+        $role->fill($request->all());
+        $role->save();
 
-
-        /* LARAVEL VALIDATION */
-        // create the validation rules
-        $rules = array(
-            'slug' => 'required|min:2|max:4',
-            'nom' => 'max:45'
-        );
-
-        $validator = Validator::make($request->all(), $rules);
-
-        if ($validator->fails() || !empty($customError)) {
-            return view('administrations.roles.edit')->withErrors($validator->errors())->with('role', $role)->with('customError', $customError);
-        } else {
-            $role->update($request->all());
-            return redirect()->route('roles.index');
-        }
+        return redirect()->route('roles.index');
     }
 
     /**
@@ -116,7 +95,11 @@ class RoleController extends Controller
     public function destroy($role)
     {
         $roleToDelete = Role::findOrFail($role);
-        $roleToDelete->delete();    
+
+        if ($roleToDelete->isUsed() == null){
+            $roleToDelete->delete();    
+        }  
+
         return redirect()->route('roles.index');
     }
 }
